Move all manual version logic together
